extract file operation / methods of the server
