+++ server/express-server.ts
@@ -27,10 +27,10 @@ import {TwitchTriggerCommand} from "../projects/contracts/src/lib/types";
 import open from 'open';
 import {Subject} from "rxjs";
 import {TAG_ROUTES} from "./rest-endpoints/tags";
-import {MediaType} from "../projects/contracts/src/lib/media.types";
+import {getFiles, mapFileInformations} from "./file.utilts";
 
-const { resolve, basename, extname, sep, normalize, join } = require('path');
-const { readdir } = fs.promises;
+const {  normalize, join } = require('path');
+
 
 var cors = require('cors')
 var bodyParser = require('body-parser');
@@ -198,35 +198,9 @@ app.put(CONFIG_TWITCH_CHANNEL_ENDPOINT, (req, res) => {
   res.send(PersistenceInstance.updateTwitchChannel(req.body.twitchChannel));
 });
 
-async function getFiles(dir) {
-  const dirents = await readdir(dir, { withFileTypes: true });
-  const files = await Promise.all(dirents.map((dirent) => {
-    const res = resolve(dir, dirent.name);
-    return dirent.isDirectory() ? getFiles(res) : res;
-  }));
-  return Array.prototype.concat(...files);
-}
 
-function fileEndingToType (fileEnding: string) : MediaType {
-  fileEnding = fileEnding.toLowerCase().replace('.', '');
 
-  switch (fileEnding) {
-    case 'jpg':
-    case 'jpeg':
-    case 'gif':
-    case 'webp':
-    case 'png':
-      return MediaType.Picture;
-    case 'mp3':
-    case 'wav':
-    case 'ogg':
-    case 'flac':
-      return MediaType.Audio;
-    case 'mp4':
-    case 'webm':
-      return MediaType.Video;
-  }
-}
+
 
 
 app.get(FILES_OPEN_ENDPOINT, async (req, res) => {
@@ -246,31 +220,13 @@ app.get(FILES_ENDPOINT, async (req, res) => {
   const files = await getFiles(mediaFolder);
 
   // files with information
-  const fileInfoList = files.map((fullPath: string) => {
-    const ext = extname(fullPath);
-    const fileName = basename(fullPath);
-
-    const fileType = fileEndingToType(ext);
-
-    // TODO replace PORT with the "--port"
-    const apiUrl = fullPath
-      .replace(mediaFolder,
-      `http://localhost:${app.get('port')}/file`
-      ).split(sep).join('/');
-
-    return {
-      fullPath,
-      fileName,
-      ext,
-      fileType,
-      apiUrl
-    }
-  });
+  const fileInfoList = mapFileInformations(mediaFolder, app.get('port'), files);
 
   res.send(fileInfoList);
 });
 
-// TODO use IDs instead of names
+// TODO use IDs instead of names ?
+// TODO use express.static ?
 // after the json "database" is done
 // use filename which is under
 // dev mode : "/src/assets"
@@ -310,6 +266,7 @@ app.get(NETWORK_IP_LIST_ENDPOINT, (req, res) => {
   // update config
   res.send(listNetworkInterfaces());
 });
+
 
 
 export function createExpress(port) {
