--- tsconfig.ts-node.json
@@ -1,40 +1,0 @@
-{
-  "compileOnSave": false,
-  "compilerOptions": {
-    "module": "commonjs",
-    "outDir": "./dist/out-tsc",
-    "sourceMap": true,
-    "baseUrl": ".",
-    "rootDir": ".",
-    "declaration": false,
-    "moduleResolution": "node",
-    "esModuleInterop": true,
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true,
-    "target": "es5",
-    "typeRoots": [
-      "node_modules/@types"
-    ],
-    "lib": [
-      "es2017",
-      "es2016",
-      "es2015",
-      "dom"
-    ],
-    "paths": {
-      "@memebox/contracts": [
-        "projects/contracts/src/public-api.ts"
-      ]
-    }
-  },
-  "files": [
-    "src/main.ts",
-    "src/polyfills.ts"
-  ],
-  "include": [
-    "src/**/*.d.ts"
-  ],
-  "exclude": [
-    "node_modules"
-  ]
-}
